pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prmkit"
version = "1.0.0"
description = "Gmail到Google Groups多任务邮件迁移工具"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "PRMKit Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["gmail", "google-groups", "email-migration", "automation"]

dependencies = [
    "flask>=3.0.0",
    "google-api-python-client==2.100.0",
    "google-auth==2.23.3",
    "google-auth-oauthlib==1.0.0",
    "google-auth-httplib2==0.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/your-username/prmkit"
Repository = "https://github.com/your-username/prmkit.git"
Issues = "https://github.com/your-username/prmkit/issues"
Documentation = "https://github.com/your-username/prmkit#readme"

[project.scripts]
prmkit = "app:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[tool.uv.sources]
# UV 特定配置
# 可以在这里配置私有包源或特定的包版本源

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["."]  # 包含当前目录的所有 Python 文件

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]